Add basic test suite
Adds basic test suite to allow for splitting unit tests and integration tests.

In the future we'll need to expand the `TestCase{}` struct with helper functions and utilities. For now, this will help us expand on our tests without fear of creating new resources everytime we want to run `make test`.