Refactor the Client Library
Refactors the client library to completely remove the authenticatedClient.

Allows us to create different clients for the terraform provider with the same OPC config struct, currently only the compute client is available.

Also creates the OPCError type to better handle errors.

This allows the provider initialization to be refactored (somewhat) into:

```
type OPCClient struct {
  computeClient *compute.Client
}

func (c *Config) Client() (*OPCClient, error) {
  var client OPCClient
  u, err := url.ParseRequestURI(c.Endpoint)
  if err != nil {
    return nil, fmt.Errorf("Invalid endpoint URI: %s", err)
  }

  conf := &opc.Config{
    IdentityDomain: &c.IdentityDomain,
    Username: &c.Username,
    Password: &c.Password,
    APIEndpoint: u
  }

  client.computeClient, err := compute.NewComputeClient(conf)
  if err != nil {
    return nil, err
  }

  return client, nil
}
```

This allows us to write multiple clients for the OPCClient type, and add those when they are created at a future date.

We desperately need to write a logger that can be used with the OPC Config, and use that logger throughout the SDK. This will allow the SDK to log at the same logging level that Terraform logs with, as we can pass Terraform's logger into the OPC Config, which will allow us to log the HTTP responses and have a better debugging output story.