Add Logger
Allows the user to add a logger to the OPC configuration. Also adds debug logging functions for sending HTTP Requests and reading HTTP Responses.

During the resource refactors we should now be able to add debug logging with the helper log functions, as well as fix the logging in each of the resource tests.