Finalize logging
After this changeset, tests will now be able to simply use the `log` package to log statements to the user, and the logs will only be displayed if the `ORACLE_LOG` environment variable is set. This allows for cleaner test output unless debug logs are desired.

This also allows the developer to write debug log statements from inside the `compute` package, by  exposing the `client.logger.Log()` interface, through the `client.debugLogString()` method. This method will only log the provided string, IF, the client's log level is set to `LogDebug`, and will use the provided `Logger` interface, whether it's the default logger or an end-user supplied logger.

 Also wrote some more documentation in the Readme for the SDK.

 Also adds a convert helper function to convert raw strings to string pointers.