Fix Storage client + Add file upload support
Fully adds file upload support, and adds API methods to send a non-JSON
request body so the file that's attached to the Request body isn't marshaled into JSON.

Adds and finishes tests for the StorageClientObject.

Adds helper API method for creating a resource with a body specified in the HTTP request.