storage_volumes: Actually capture errors during a storage volume creation
Previously we were only capturing errors returned from the API, not any
errors on Storage Volume creation, thus causing the storage volume to sit
in a wait-ready loop until we hit a timeout, even if the storage volume
had failed to create a long time ago.

This checks if the returned storage volume's status contains `"error"`,
and returns the status detailed text, if so.